增加判断文章是否存在在删除文章前，防止空指针

+++ src/main/java/com/ikaileblog/service/impl/ArticleServiceImpl.java
@@ -185,6 +185,10 @@ public class ArticleServiceImpl implements ArticleService {
         // article本身
         // articleBody
         // article_tag
+        QueryWrapper<Article> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("id", id);
+        Article article = articleMapper.selectOne(queryWrapper);
+        if (article == null) return new RestBean<>(404, "文章不存在");
         articleMapper.deleteById(id);
         articleBodyMapper.deleteArticleBodyById(id);
         articleTagMapper.deleteArticleTagById(id);
